Cleanup

+++ src/main/kotlin/me/honnold/ladderhero/repository/PlayerRepository.kt
@@ -1,6 +1,6 @@
 package me.honnold.ladderhero.repository
 
-import me.honnold.ladderhero.model.db.Player
+import me.honnold.ladderhero.domain.Player
 import org.springframework.data.r2dbc.repository.Query
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Mono
@@ -8,8 +8,8 @@ import java.util.*
 
 interface PlayerRepository : ReactiveCrudRepository<Player, UUID> {
     @Query("select count(*) > 0 from player where profile_id = :profileId")
-    fun existsByProfileId(profileId: Int): Mono<Boolean>
+    fun existsByProfileId(profileId: Long): Mono<Boolean>
 
     @Query("select * from player where profile_id = :profileId")
-    fun findByProfileId(profileId: Int): Mono<Player>
+    fun findByProfileId(profileId: Long): Mono<Player>
 }
